Removed legacy components, moved get_llm_instance to app_qa_builder
