+++ scripts/app_qa_builder.py
@@ -2,6 +2,7 @@ import os
 import textwrap
 from typing import Optional
 from urllib.request import pathname2url
+import logging
 
 from deep_translator import GoogleTranslator
 from langchain import PromptTemplate
@@ -9,34 +10,132 @@ from langchain.chains import ConversationalRetrievalChain
 from langchain.chains.retrieval_qa.base import BaseRetrievalQA
 from langchain.embeddings import OpenAIEmbeddings, HuggingFaceEmbeddings
 from langchain.vectorstores import Chroma
+from langchain.callbacks.base import BaseCallbackHandler
+from langchain import HuggingFacePipeline
+
+from langchain.llms import LlamaCpp, GPT4All, OpenAI
+from transformers import AutoTokenizer, AutoModelForCausalLM, LlamaTokenizer, LlamaForCausalLM, GenerationConfig, pipeline
 from openai.error import AuthenticationError
 
 from .app_environment import translate_dst, translate_src, translate_docs, translate_q, ingest_target_source_chunks, args, openai_use, ingest_embeddings_model, gpu_is_enabled, \
     chromaDB_manager
 
-
-def print_hyperlink(doc):
-    page_link = doc.metadata['source']
-    abs_path = os.path.abspath(page_link)
-    file_url = pathname2url(abs_path)
-
-    # This is the URL-encoded path, which can be used in a browser or link
-    print(f'\033[32m[!]\033[0m URL: file://{file_url}')
-
-    # This is the original path, which might contain characters not allowed in URLs (like spaces)
-    print(f'\033[32m[!]\033[0m Path: {page_link}')
+from .app_environment import model_type, openai_api_key, model_n_ctx, model_temperature, model_top_p, model_n_batch, model_use_mlock, model_verbose, \
+    db_get_only_relevant_docs, gpt4all_backend, model_path_or_id, cpu_model_n_threads, gpu_model_n_threads, model_n_answer_words, huggingface_model_base_name
 
 
-def print_document_chunk(doc):
-    document_page = doc.page_content.replace('\n', ' ')
-    if translate_docs:
-        document_page = GoogleTranslator(source=translate_src, target=translate_dst).translate(document_page)
-    wrapper = textwrap.TextWrapper(initial_indent='\033[37m', subsequent_indent='\033[37m', width=120)
-    print(f"{wrapper.fill(document_page)}\033[0m\n")
-    print(f'\033[94m"n" -> next, "q" -> quit: \033[0m')
-    user_input = input()
-    if user_input.lower() == 'q':
-        exit(0)
+# Ensure TOKENIZERS_PARALLELISM is set before importing any HuggingFace module.
+os.environ["TOKENIZERS_PARALLELISM"] = "true"
+# load environment variables
+
+try:
+    load_dotenv()
+except Exception as e:
+    logging.error("Error loading .env file, create one from example.env:", str(e))
+
+
+# noinspection PyPep8Naming
+def calculate_layer_count() -> None | int | float:
+    """
+    How many layers of a neural network model you can fit into the GPU memory,
+    rather than determining the number of threads.
+    The layer size is specified as a constant (120.6 MB), and the available GPU memory is divided by this to determine the maximum number of layers that can be fit onto the GPU.
+    Some additional memory (the size of 6 layers) is reserved for other uses.
+    The maximum layer count is capped at 32.
+    """
+    if not gpu_is_enabled:
+        return None
+    LAYER_SIZE_MB = 120.6  # This is the size of a single layer on VRAM, and is an approximation.
+    # The current set value is for 7B models. For other models, this value should be changed.
+    LAYERS_TO_REDUCE = 6  # About 700 MB is needed for the LLM to run, so we reduce the layer count by 6 to be safe.
+    if (get_gpu_memory() // LAYER_SIZE_MB) - LAYERS_TO_REDUCE > 32:
+        return 32
+    else:
+        return get_gpu_memory() // LAYER_SIZE_MB - LAYERS_TO_REDUCE
+
+
+def get_llm_instance(*callback_handler: BaseCallbackHandler):
+    logging.debug(f"Initializing model...")
+
+    callbacks = [] if args.mute_stream else callback_handler
+
+    if model_type == "gpt4all":
+        if gpu_is_enabled:
+            logging.warn("GPU is enabled, but GPT4All does not support GPU acceleration. Please use LlamaCpp instead.")
+            exit(1)
+        return GPT4All(
+            model=model_path_or_id,
+            n_ctx=model_n_ctx,
+            backend=gpt4all_backend,
+            callbacks=callbacks,
+            use_mlock=model_use_mlock,
+            n_threads=gpu_model_n_threads if gpu_is_enabled else cpu_model_n_threads,
+            n_predict=1000,
+            n_batch=model_n_batch,
+            top_p=model_top_p,
+            temp=model_temperature,
+            streaming=False,
+            verbose=False
+        )
+    elif model_type == "llamacpp":
+        return LlamaCpp(
+            model_path=model_path_or_id,
+            temperature=model_temperature,
+            n_ctx=model_n_ctx,
+            top_p=model_top_p,
+            n_batch=model_n_batch,
+            use_mlock=model_use_mlock,
+            n_threads=gpu_model_n_threads if gpu_is_enabled else cpu_model_n_threads,
+            verbose=model_verbose,
+            n_gpu_layers=calculate_layer_count() if gpu_is_enabled else None,
+            callbacks=callbacks,
+        )
+    elif model_type == "huggingface":
+        if gpu_is_enabled and huggingface_model_base_name is not None:
+            logging.info("Tokenizer loaded")
+            tokenizer = AutoTokenizer.from_pretrained(model_path_or_id, use_fast=True)
+            model = AutoGPTQForCausalLM.from_quantized(
+                model_name_or_path=model_path_or_id,
+                model_basename=huggingface_model_base_name if ".safetensors" not in huggingface_model_base_name else huggingface_model_base_name.replace(".safetensors", ""),
+                use_safetensors=True,
+                trust_remote_code=True,
+                device="cuda:0",
+                use_triton=False,
+                quantize_config=None,
+            )
+        elif gpu_is_enabled:
+            logging.info("Using AutoModelForCausalLM for full models")
+            tokenizer = AutoTokenizer.from_pretrained(model_path_or_id)
+            model = AutoModelForCausalLM.from_pretrained(
+                model_path_or_id,
+                device_map="auto",
+                torch_dtype=torch.float16,
+                low_cpu_mem_usage=True,
+                trust_remote_code=True,
+                # max_memory={0: "15GB"} # Uncomment this line if you encounter CUDA out of memory errors
+            )
+            model.tie_weights()
+        else:
+            logging.info("Using LlamaTokenizer")
+            tokenizer = LlamaTokenizer.from_pretrained(model_path_or_id)
+            model = LlamaForCausalLM.from_pretrained(model_path_or_id)
+
+        return HuggingFacePipeline(pipeline=pipeline(
+            "text-generation",
+            model=model,
+            tokenizer=tokenizer,
+            max_length=2048,
+            temperature=0,
+            top_p=model_top_p,
+            repetition_penalty=1.15,
+            generation_config=GenerationConfig.from_pretrained(model_path_or_id),
+        ))
+    elif model_type == "openai":
+        assert openai_api_key is not None, "Set ENV OPENAI_API_KEY, Get one here: https://platform.openai.com/account/api-keys"
+        return OpenAI(openai_api_key=openai_api_key, callbacks=callbacks)
+    else:
+        logging.error(f"Model {model_type} not supported!")
+        raise Exception(f"Model type {model_type} is not supported. Please choose one of the following: LlamaCpp, GPT4All")
 
 
 async def process_database_question(database_name, llm, collection_name: Optional[str]):
